fix class register problem
